backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sqlmodel==0.0.8
passlib[bcrypt]==1.7.4
pyjwt==2.8.0
python-multipart==0.0.6
opencv-python-headless==4.7.0.72
numpy==1.26.4
pillow==10.1.0
redis==5.0.4
rq==1.13.1
aiofiles==23.1.0
python-dotenv==1.0.1
